feat: implement cjsToEsm.cjs
